% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phase.plot.R
\name{arrow}
\alias{arrow}
\title{Helper function for phase.plot}
\usage{
arrow(x, y, l = 0.1, w = 0.3 * l, alpha, col = "black")
}
\arguments{
\item{x}{TODO}

\item{y}{TODO}

\item{l}{TODO}

\item{w}{TODO}

\item{alpha}{TODO}

\item{col}{TODO}
}
\description{
Helper function for phase.plot
}

