% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase.plot.R
\name{arrow2}
\alias{arrow2}
\title{This is an alternative helper function that plots arrows.
It uses text() to print a character using a default font.
This way, it is possible to render different types of arrows.}
\usage{
arrow2(x, y, angle, size = 0.1, col = "black", chr = "➛")
}
\arguments{
\item{x}{X-coordinate of the arrow.}

\item{y}{Y-coordinate of the arrow.}

\item{angle}{Angle in radians}

\item{size}{Similar to \code{arrow.len} parameter. Notice that we don't need
the arrow.lwd anymore}

\item{col}{Color of the arrow.}

\item{chr}{Character representing the arrow. You should provide the character
as escaped UTF-8.}
}
\description{
This is an alternative helper function that plots arrows.
It uses text() to print a character using a default font.
This way, it is possible to render different types of arrows.
}
\examples{
# Not run: arrow2(x[j], y[i], angle = phases[i, j],
# Not run:        col = arrow.col, size = arrow.len)
}
