% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/smooth.wavelet.R
\name{boxcar_scale_smoothing}
\alias{boxcar_scale_smoothing}
\title{Scale smoothing (boxcar with width of 0.6)}
\usage{
boxcar_scale_smoothing(twave, dj)
}
\arguments{
\item{twave}{matrix}

\item{dj}{number of octaves per scale}
}
\value{
swave
}
\description{
This helper function will be later reimplemented in c++
}

