% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.data.R
\name{check.data}
\alias{check.data}
\title{Check the format of time series}
\usage{
check.data(y, x1 = NULL, x2 = NULL)
}
\arguments{
\item{y}{Time series \code{y} in matrix format (\code{n} rows x 2 columns).
The first column should contain the time steps and the second column should
contain the values.}

\item{x1}{Time series \code{x1} in matrix format (\code{n} rows x 2 columns).
The first column should contain the time steps and the second column should
contain the values.}

\item{x2}{Time series \code{x2} in matrix format (\code{n} rows x 2 columns).
The first column should contain the time steps and the second column should
contain the values.}
}
\value{
Returns a named list containing:
\item{t}{Time steps}
\item{dt}{Size of a time step}
\item{n.obs}{Number of observations}
}
\description{
Check the format of time series
}
\examples{
t1 <- cbind(1:100, rnorm(100))
check.data(y = t1)

}
\references{
Torrence, C., and G. P. Compo. 1998. A Practical Guide to Wavelet Analysis.
\emph{Bulletin of the American Meteorological Society} 79:61-78.
}
\author{
Tarik C. Gouhier (tarik.gouhier@gmail.com)
}
