% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtc.sig.R
\name{get_minroots}
\alias{get_minroots}
\title{Helper function (not exported)}
\usage{
get_minroots(ar)
}
\arguments{
\item{ar}{The 'ar' part of AR(1)}
}
\value{
double
}
\description{
Helper function (not exported)
}
