% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biwavelet-package.R
\docType{data}
\name{MOTHERS}
\alias{MOTHERS}
\title{Supported mother wavelets}
\format{\preformatted{ chr [1:3] "morlet" "paul" "dog"
}}
\usage{
MOTHERS
}
\description{
The list of supported mother wavelets is used in multiple places
therefore, we provide it as a lazily evaluated promise.
}
\keyword{datasets}

