% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_wt_bases_dog}
\alias{rcpp_wt_bases_dog}
\title{Optimized "wt.bases.dog" function.}
\usage{
rcpp_wt_bases_dog(k, scale, param = -1L)
}
\arguments{
\item{k}{vector of frequencies at which to calculate the wavelet.}

\item{scale}{the wavelet scale.}

\item{param}{nondimensional parameter specific to the wavelet function.}
}
\value{
Returns a list containing:
\item{daughter}{wavelet function}
\item{fourier.factor}{ratio of fourier period to scale}
\item{coi}{cone of influence}
\item{dof}{degrees of freedom for each point in wavelet power}
}
\description{
This is a C++ version optimized for speed.
Computes the wavelet as a function of Fourier frequency
for "dog" mother wavelet.
}
\note{
This c++ implementation is approx. 50% faster than the original R code
}
\author{
Viliam Simko
}
