% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowQuantileC.R
\name{rowQuantileC}
\alias{rowQuantileC}
\title{This function calls the C++ implementation of row quantile.}
\usage{
rowQuantileC(data, p)
}
\arguments{
\item{data}{Numeric matrix whose row quantiles are wanted.}

\item{p}{Probability with value in [0,1]}
}
\value{
A vector of length nrows(data), where each element represents row
  quantile.
}
\description{
This function calls the C++ implementation of row quantile.
}

