% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt.bases.R
\name{wt.bases.dog}
\alias{wt.bases.dog}
\title{Helper method (not exported)}
\usage{
wt.bases.dog(k, scale, param = -1)
}
\arguments{
\item{k}{Vector of frequencies at which to calculate the wavelet.}

\item{scale}{The wavelet scale.}

\item{param}{Nondimensional parameter specific to the wavelet function.}
}
\value{
Returns a list containing:
\item{daughter}{wavelet function}
\item{fourier.factor}{ratio of fourier period to scale}
\item{coi}{cone of influence}
\item{dof}{degrees of freedom for each point in wavelet power}
}
\description{
Helper method (not exported)
}
