% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolve2D.R
\name{convolve2D_typeopen}
\alias{convolve2D_typeopen}
\title{Speed-optimized version of convolve2D}
\usage{
convolve2D_typeopen(x, y)
}
\arguments{
\item{x}{M \code{x} n matrix.}

\item{y}{Numeric sequence of length N.}
}
\description{
Equivalent to \code{convolve2D(x, y, type = "open")}. The motivation for this
function was that convolution is called many times in a loop from
\code{\link{smooth.wavelet}}, always with the \code{type = "open"} parameter.
}
\seealso{
\code{\link{convolve2D}}
}
\author{
Viliam Simko
}
