% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meshgrid.R
\name{meshgrid}
\alias{meshgrid}
\title{Rectangular grid in two dimensions}
\usage{
meshgrid(xv, yv)
}
\arguments{
\item{xv}{vector of numeric values}

\item{yv}{vector of numeric values}
}
\value{
Returns a list containing the full grid components \code{xv} and \code{yv}:
\item{x }{replicated values of xv}
\item{y }{replicated values of yv}
}
\description{
Replicates the grid vectors \code{xv} and \code{yv} to generate a full grid
}
\examples{
xv <- runif(10)
yv <- runif(5)
g <- meshgrid(xv, yv)
}
\author{
Tarik C. Gouhier (tarik.gouhier@gmail.com)
}

