% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biwavelet-package.R
\docType{data}
\name{MOTHERS}
\alias{MOTHERS}
\title{Supported mother wavelets}
\format{An object of class \code{character} of length 3.}
\usage{
MOTHERS
}
\description{
The list of supported mother wavelets is used in multiple places
therefore, we provide it as a lazily evaluated promise.
}
\keyword{datasets}
